Deliver more meaningful error responses
The application should render responses to some of the more obvious
and predictable classes of error conditions that include some basic
information about the exception and a displayable error message
(or messages) with specific information to help the client correct
the problem. Use standard Rails hooks and patterns to systematize
these responses.